# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CalendarEvent {
  id: string;
  summary: string;
  start: string;
  end: string;
  allDay: boolean;
  /** @nullable */
  htmlLink: string | null;
}

export interface CalendarSummary {
  currentDate: string;
  month: string;
  timeZone: string;
  events: CalendarEvent[];
}

export interface ErrorResponse {
  error: string;
}

export type GetCalendarSummaryParams = {
/**
 * IANA timezone used for the month window and date labels.
 */
timeZone?: string;
};