# lib/api-zod/src/generated/types/calendarSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarEvent } from './calendarEvent';

export interface CalendarSummary {
  currentDate: string;
  month: string;
  timeZone: string;
  events: CalendarEvent[];
}